requirements.txt
streamlit==1.45.1
langchain==0.3.25
langchain-deepseek==0.1.3
langchain-mcp-adapters==0.0.10
langgraph==0.4.5
langgraph-prebuilt==0.1.8
pydantic==2.10.6
python-dotenv==1.1.0
typing-extensions==4.13.2
asyncio==3.4.3
langchain-core==0.3.59
langchain-community==0.3.24
